PasswordProtectedFiles

+++ code/modules/modular_computers/file_system/programs/system/file_browser.dm
@@ -19,7 +19,18 @@
 
 	if(href_list["PRG_openfile"])
 		. = 1
-		open_file = href_list["PRG_openfile"]
+		var/obj/item/weapon/computer_hardware/hard_drive/HDD = computer.hard_drive
+		var/datum/computer_file/file = HDD.find_file_by_name(href_list["PRG_openfile"])
+		var/checkpass = ""
+		var/currentpass = file.password
+		if(!currentpass)
+			open_file = href_list["PRG_openfile"]
+		if(currentpass)
+			checkpass = sanitize(input(usr, "Please enter the password:"))
+			if(checkpass == currentpass)
+				open_file = href_list["PRG_openfile"]
+			if(!checkpass == currentpass)
+				return 1
 	if(href_list["PRG_newtextfile"])
 		. = 1
 		var/newname = sanitize(input(usr, "Enter file name or leave blank to cancel:", "File rename"))
@@ -28,9 +39,11 @@
 		var/obj/item/weapon/computer_hardware/hard_drive/HDD = computer.hard_drive
 		if(!HDD)
 			return 1
+		var/newpassword = sanitize(input(usr, "Enter a password or leave blank to leave unprotected:"))
 		var/datum/computer_file/data/F = new/datum/computer_file/data()
 		F.filename = newname
 		F.filetype = "TXT"
+		F.password = newpassword
 		HDD.store_file(F)
 	if(href_list["PRG_deletefile"])
 		. = 1
@@ -38,18 +51,36 @@
 		if(!HDD)
 			return 1
 		var/datum/computer_file/file = HDD.find_file_by_name(href_list["PRG_deletefile"])
+		var/checkpass = ""
+		var/currentpass = file.password
 		if(!file || file.undeletable)
 			return 1
-		HDD.remove_file(file)
+		if(!currentpass)
+			HDD.remove_file(file)
+		if(currentpass)
+			checkpass = sanitize(input(usr, "Please enter the password:"))
+			if(checkpass == currentpass)
+				HDD.remove_file(file)
+			if(!checkpass == currentpass)
+				return 1
 	if(href_list["PRG_usbdeletefile"])
 		. = 1
 		var/obj/item/weapon/computer_hardware/hard_drive/RHDD = computer.portable_drive
 		if(!RHDD)
 			return 1
 		var/datum/computer_file/file = RHDD.find_file_by_name(href_list["PRG_usbdeletefile"])
+		var/checkpass = ""
+		var/currentpass = file.password
 		if(!file || file.undeletable)
 			return 1
-		RHDD.remove_file(file)
+		if(!currentpass)
+			RHDD.remove_file(file)
+		if(currentpass)
+			checkpass = sanitize(input(usr, "Please enter the password:"))
+			if(checkpass == currentpass)
+				RHDD.remove_file(file)
+			if(!checkpass == currentpass)
+				return 1
 	if(href_list["PRG_closefile"])
 		. = 1
 		open_file = null
@@ -60,21 +91,40 @@
 		if(!HDD)
 			return 1
 		var/datum/computer_file/F = HDD.find_file_by_name(href_list["PRG_clone"])
+		var/checkpass = ""
+		var/currentpass = F.password
 		if(!F || !istype(F))
 			return 1
 		var/datum/computer_file/C = F.clone(1)
-		HDD.store_file(C)
+		if(!currentpass)
+			HDD.store_file(C)
+		if(currentpass)
+			checkpass = sanitize(input(usr, "Please enter the password:"))
+			if(checkpass == currentpass)
+				HDD.store_file(C)
+			if(!checkpass == currentpass)
+				return 1
 	if(href_list["PRG_rename"])
 		. = 1
 		var/obj/item/weapon/computer_hardware/hard_drive/HDD = computer.hard_drive
 		if(!HDD)
 			return 1
 		var/datum/computer_file/file = HDD.find_file_by_name(href_list["PRG_rename"])
+		var/checkpass = ""
+		var/currentpass = file.password
 		if(!file || !istype(file))
 			return 1
 		var/newname = sanitize(input(usr, "Enter new file name:", "File rename", file.filename))
-		if(file && newname)
-			file.filename = newname
+		if(!currentpass)
+			if(file && newname)
+				file.filename = newname
+		if(currentpass)
+			checkpass = sanitize(input(usr, "Please enter the password:"))
+			if(checkpass == currentpass)
+				if(file && newname)
+					file.filename = newname
+			if(!checkpass == currentpass)
+				return 1
 	if(href_list["PRG_edit"])
 		. = 1
 		if(!open_file)
